fixed minfloat and build issues Co-authored-by: redfish-bluefish <redfish-bluefish@users.noreply.github.com> Co-authored-by: Dawson Whitehead <dwahme@users.noreply.github.com> Co-authored-by: Lily Ehsani <lilyehsani@users.noreply.github.com>

+++ src/minifloat/src/minifloat.rs
@@ -14,7 +14,7 @@ impl Minifloat {
         }
 
         Minifloat {
-            value: (f * 8) as i8
+            value: (f * 8 as f32) as i8
         }
     }
